[ADD] Tile: count_drone & count_item. Tested
[ADD] Incantation: create, destroy, check. Tested

+++ server/src/simulation/incantation/find_requirements.c
@@ -7,7 +7,7 @@
 
 #include "simulation/incantation.h"
 
-const elevation_requirement_t *incantation_find_requirements(size_t drone_lvl)
+const elevation_requirement_t *incantation_find_requirements(int drone_lvl)
 {
     for (size_t i = 0; i < elevation_requirements_size; i++) {
         if (elevation_requirements[i].drone_lvl == drone_lvl)
